Detect Chromium during remote boot
