+++ packages/playground/client/src/blueprints-v1-handler.spec.ts
@@ -16,7 +16,6 @@ const mocks = vi.hoisted(() => {
 		runBlueprintV1Steps: vi.fn(),
 		resolveRuntimeConfiguration: vi.fn(),
 		createBlueprintReflection: vi.fn(),
-		isChromiumBasedBrowser: vi.fn(),
 		consumeAPI: vi.fn(),
 		collectPhpLogs: vi.fn(),
 	};
@@ -29,10 +28,6 @@ vi.mock('@php-wasm/logger', () => ({
 
 vi.mock('@php-wasm/universal', () => ({
 	consumeAPI: mocks.consumeAPI,
-}));
-
-vi.mock('./browser-detection', () => ({
-	isChromiumBasedBrowser: mocks.isChromiumBasedBrowser,
 }));
 
 vi.mock('.', () => ({
@@ -55,7 +50,6 @@ describe('BlueprintsV1Handler', () => {
 		mocks.playground.prefetchUpdateChecks.mockResolvedValue(undefined);
 		mocks.compileBlueprintV1.mockResolvedValue([]);
 		mocks.isBlueprintBundle.mockReturnValue(false);
-		mocks.isChromiumBasedBrowser.mockReturnValue(false);
 		mocks.runBlueprintV1Steps.mockResolvedValue(undefined);
 		mocks.resolveRuntimeConfiguration.mockResolvedValue({
 			phpVersion: '8.4',
@@ -119,24 +113,6 @@ describe('BlueprintsV1Handler', () => {
 			})
 		);
 		expect(progressTracker.pipe).toHaveBeenCalledWith(mocks.playground);
-	});
-
-	it('passes Chromium admin View Transition detection to the runtime', async () => {
-		mocks.isChromiumBasedBrowser.mockReturnValue(true);
-		const iframe = createIframe();
-		const handler = new BlueprintsV1Handler({
-			iframe,
-			remoteUrl: 'http://example.com/remote.html',
-			blueprint: {},
-		});
-
-		await handler.bootPlayground(iframe, createProgressTracker());
-
-		expect(mocks.playground.boot).toHaveBeenCalledWith(
-			expect.objectContaining({
-				disableAdminViewTransitions: true,
-			})
-		);
 	});
 
 	it('does not pipe progress to the remote when progress bar is disabled', async () => {
