+++ packages/playground/client/src/blueprints-v2-handler.ts
@@ -2,7 +2,6 @@ import type { ProgressTracker } from '@php-wasm/progress';
 import type { PlaygroundClient, StartPlaygroundOptions } from '.';
 import { collectPhpLogs, logger } from '@php-wasm/logger';
 import { consumeAPI } from '@php-wasm/universal';
-import { isChromiumBasedBrowser } from './browser-detection';
 
 export class BlueprintsV2Handler {
 	private readonly options: StartPlaygroundOptions;
@@ -102,7 +101,6 @@ export class BlueprintsV2Handler {
 			// Pass the declaration directly – the worker runs the V2 runner.
 			blueprint: blueprint as any,
 			pathAliases,
-			disableAdminViewTransitions: isChromiumBasedBrowser(),
 		} as any);
 
 		await playground.isReady();
