+++ packages/playground/remote/src/lib/playground-worker-endpoint-blueprints-v1.ts
@@ -44,7 +44,6 @@ class PlaygroundWorkerEndpointBlueprintsV1 extends PlaygroundWorkerEndpoint {
 		wordpressInstallMode,
 		corsProxyUrl,
 		pathAliases,
-		disableAdminViewTransitions,
 	}: WorkerBootOptions) {
 		if (this.booted) {
 			throw new Error('Playground already booted');
@@ -75,7 +74,6 @@ class PlaygroundWorkerEndpointBlueprintsV1 extends PlaygroundWorkerEndpoint {
 				withNetworking,
 				phpVersion: phpVersion!,
 				pathAliases,
-				disableAdminViewTransitions,
 			});
 
 			this.requestedWordPressVersion =
