+++ packages/playground/remote/src/lib/playground-worker-endpoint-blueprints-v2.ts
@@ -26,7 +26,6 @@ class PlaygroundWorkerEndpointV2 extends PlaygroundWorkerEndpoint {
 		corsProxyUrl,
 		blueprint,
 		pathAliases,
-		disableAdminViewTransitions,
 	}: WorkerBootOptions) {
 		if (this.booted) {
 			throw new Error('Playground already booted');
@@ -50,7 +49,6 @@ class PlaygroundWorkerEndpointV2 extends PlaygroundWorkerEndpoint {
 				withNetworking,
 				phpVersion: phpVersion!,
 				pathAliases,
-				disableAdminViewTransitions,
 			});
 			const primaryPhp = await requestHandler.getPrimaryPhp();
