+++ packages/playground/remote/src/lib/playground-worker-endpoint.ts
@@ -94,10 +94,6 @@ export type WorkerBootOptions = {
 	 * the document root. Similar to Nginx's `alias` directive.
 	 */
 	pathAliases?: PathAlias[];
-	/**
-	 * Disables browser-native View Transitions for wp-admin pages.
-	 */
-	disableAdminViewTransitions?: boolean;
 };
 
 /** @inheritDoc PHPClient */
@@ -156,7 +152,6 @@ export abstract class PlaygroundWorkerEndpoint extends PHPWorker {
 		withNetworking,
 		phpVersion,
 		pathAliases,
-		disableAdminViewTransitions,
 	}: {
 		siteUrl: string;
 		sapiName: string;
@@ -166,7 +161,6 @@ export abstract class PlaygroundWorkerEndpoint extends PHPWorker {
 		withNetworking: boolean;
 		phpVersion: AllPHPVersion;
 		pathAliases?: PathAlias[];
-		disableAdminViewTransitions?: boolean;
 	}) {
 		const phpIniEntries: Record<string, string> = {
 			'openssl.cafile': '/internal/shared/ca-bundle.crt',
@@ -278,9 +272,7 @@ export abstract class PlaygroundWorkerEndpoint extends PHPWorker {
 				'/internal/shared/ca-bundle.crt': caBundleContent,
 				'/internal/shared/mu-plugins': {
 					'0-playground-runtime-options.php':
-						createRuntimeOptionsMuPlugin({
-							disableAdminViewTransitions,
-						}),
+						createRuntimeOptionsMuPlugin(),
 					// Legacy PHP can't parse closures at all (even with an
 					// early return), so use a minimal compatible stub instead.
 					'1-playground-web.php': isLegacyPhp
@@ -582,19 +574,33 @@ export abstract class PlaygroundWorkerEndpoint extends PHPWorker {
 	}
 }
 
-function createRuntimeOptionsMuPlugin({
-	disableAdminViewTransitions,
-}: {
-	disableAdminViewTransitions?: boolean;
-}) {
+function createRuntimeOptionsMuPlugin() {
 	const definitions: string[] = [];
-	if (disableAdminViewTransitions) {
+	if (isChromiumBasedBrowser()) {
 		definitions.push(
 			"define('PLAYGROUND_DISABLE_ADMIN_VIEW_TRANSITIONS', true);"
 		);
 	}
 
 	return `<?php\n${definitions.join('\n')}\n`;
+}
+
+function isChromiumBasedBrowser(navigatorObject: Navigator = navigator) {
+	const brands = (
+		navigatorObject as Navigator & {
+			userAgentData?: { brands?: Array<{ brand: string }> };
+		}
+	).userAgentData?.brands;
+
+	if (brands) {
+		return brands.some(({ brand }) =>
+			['Chromium', 'Google Chrome', 'Microsoft Edge', 'Opera'].includes(
+				brand
+			)
+		);
+	}
+
+	return /\b(?:Chrome|Chromium|Edg|OPR)\//.test(navigatorObject.userAgent);
 }
 
 function createNullPrototypeRecord<T>() {
